Replace winget `useLatest: true` w/ `upgrade --all`
Running the upgrades as part of `winget configure` was excruciating due
to the constant RPC failures and general slowness of winget and DSC.
This hopefully decreases the chances of failure and speeds up the
process. However, it may still be desirable to move winget upgrade to
the end of chezmoi apply (if possible).